backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
channels==4.2.2
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
dataclasses==0.6
Django==4.1.13
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-requests==0.0.2
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djongo==1.3.6
dnspython==2.7.0
idna==3.10
kombu==5.5.2
numpy==2.2.4
prompt_toolkit==3.0.50
pymongo==3.12.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sqlparse==0.2.4
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
